IPTV/HLS/M3U Remove check on m3u substring in url
Remove the check on the presence of m3u or m3u8 at the end of the url
to determine if a file is a HLS file or not.
This check prevents some IPTV channels to be played with MythTV.
There is already a check on the content of the file pointed to by
the url so the check on the url can safely be removed.

Refs #936