UDP crash after change by clang-tidy
The change of an UDP packet copy to a const reference in
rtp/packetbuffer.cpp by clang-tidy does cause a crash of the backend.
This change is from commit 816a46d13ac7267a9fbacdc2131ccb4cf68f88a1
as found with a bisect.
Reverting the change in rtp/packetbuffer.cpp.

Refs #955