IPTV/HLS/M3U Playlist reload interval
Compute the playlist reload interval as a fraction or a multiple
of the EXT-X-TARGETDURATION time. Usually this is 0.5 times the
target duration time, depending on download timing.
The logic for this has not changed, but due to a bug in the code
the result of the computation was always 0 seconds, resulting in
a reload interval time of 1 second independent of the target
duration. This bug is now fixed in this commit.

Refs #936