Default channel number for IPTV channel scan
When the IPTV channel scan does not contain a channel number for a
channel then a default channel number is used. This is the first
unused channel number in the database for that video source.
The code for this looks at all channels, but the channels that
have been deleted should be excluded. This is now fixed.

Refs #936
Refs #973