IPTV channel scan with existing channels
When a channel found in an IPTV channel scan does not have
a channel number but only a name then it does now look in
the database for the channel number.
The lookup is done on channel name only.
The lookup is done in the channels of the video source
attached to the capture card used in the channel scan.
When there is no channel number found in the IPTV channel
scan and also not in the database then a default channel
number is used. This is the first unused channel number in
the database for that video source.

Refs #936
Refs #973