Delete IPTV channel data when the channel is removed
Delete all IPTV channel data extensions records from
the database when the channel it refers to does not
exist anymore.
As part of the periodic housekeeping tasks all channels
that have been marked for deletion and for which there are
no more recordings preseent are removed from the database.
For IPTV channels the IPTV URL is stored in table iptv_channel.
This table can be seen as an extension of table channel and
it is usually accessed with the chanid from the channel table.
This means that when a channel is really deleted (removed from
the database) also the corresponding record in iptv_channel
has to be deleted.

+    // Delete all IPTV channel data extension records from the database
+    // when the channel it refers to does not exist anymore.