Add index on chanid in iptv_channel and channelgroup
The tables iptv_channel and channelgroup are accessed via
the chanid but there was no index on chanid in both tables.
Having these indices greatly speeds up the start of playback,
both for recordings and for live TV, when there is a large
number of channels available in various video sources.
This is influenced by the automatic creation of
channelgroups per video source when there is more than
one video source present in the system.