Remove duplicate browse channel list (#989)
When starting TV playback, either from a recording or live, a list
of all channels is created that can be used to select another
channel with the up/down keys while watching.
This channel list is created twice; as the name implies one
for all channels and one for all visible channels.
However, both lists are identical and contain all channels.
The check on visible or not was already been done somewhere
else in the code.
Therefore only one list with all channels is needed.
Note that the time to create this list can be significant when
there are a large number of channels. With satellites and especially
with IPTV the number of channels can easily become very large,
as in several thousands.