Update for Qt6.5 changes to QDomDocument.
The big change here is to setContent(), where the result code and all
the output arguments have been combined into a single object.  That
object can be evaluated in a boolean context to get the success or
failure of the function call, or the individual fields containing the
error information can be referenced.

https://doc.qt.io/qt-6/qdomdocument.html#setContent