Tweak initial creation of the "mythconverg" database. (2)
This change removes the display widths from all numbers when
specifying field types values in a CREATE TABLE statement.  This
matches the current output of mysqldump against a mysql database.
(Mariadb currently still includes these values.)  The mysql
documentation explicitly states that these values are deprecated and
support for them may be removed in a future release.  The only
exception to this wholesale removal of display widths is the signed
version of "tinyint(1)" which mysql keeps intact.  (Mysql keeps the
signed version of this, but converts the unsigned version from
"tinyint(1) unsigned" to "tinyint unsigned".)

These changes are best viewed with the "--word-diff" argument to the
"git diff" command.