Upgrade initial creation of the "mythconverg" database to version 1382.
Upgrade the initial database creation code to generate version 1382 of
the database (instead of version 1307).  This was done by deleting the
mythconverg database from mysql, creating a new instance of the
initial database using the command "mythbackend --upgradedbonly", and
then dumping its contents with mysqldump.  The InitializeMythSchema
function was then updated to match the output of mysqldump.  This was
done by deleting the entire contents of the "DBUpdates updates {}"
array and replacing it with the output of mysqldump.

The changes in the mysqldump invocation (the sed commands) are to
remove utf8 multibyte version specific references and replace them
with generic utf8 references (e.g. utf8mb3 -> utf8).  Mysql/mariadb
will use its default multibyte support (three or four bytes) when it
encounters these generic utf8 declarations.

Note: The replacement SQL is from mysqldump against a mysql database.
It has been loaded into both mysql and mariadb and produces the same
results as the prior code that created 1307 and upgraded to 1382.