IPTV Download not OK with network error (#999)
* IPTV Download not OK with network error

Set download status to not OK when a network error is reported
for the download. This fixes an endless loop / log pollution when
server access is not allowed, reported as  error 201 (HTTP error 403).

* IPTV Use PROGRAM-ID value 1 if not specified

The PROGRAM-ID attribute of the EXT-X-STREAM-INF is removed in
version 6 of the protocol as described in rfc8216.
Searching through a large number of M3U8 files shows that the
PROGRAM-ID value is always 1 when specified. Therefore, if the
attribute is not specified then value 1 is used instead of -1.