cmake: Notice if the install directory has been deleted.
The cmake process for building the MythTV sub-project expects to find
FFmpeg executables already in the install directory.  It runs one of
these programs to determine available features.  If the program is no
longer present, complain and stop instead of continuing with
incomplete information.

The results of find_xxx are cached, so if the program has been deleted
since the first run, find_xx stills return the location of where it
used to be.  The new "if" test is executed every time to see if the
program is actually still there.