Tell clang to ignore shadowing of enum names.
The enum name QPSK is defined in both of these header files and also
in the system header file dvb/frontend.h.  Instead of trying to fins a
better name (there isn't one) and replacing all uses of this value,
disable the compiler "shadow" warning long enough to define this enum.