MythCodecMap::GetCodecContext(): simplify if logic
If (Stream == nullptr || Stream->codecpar == nullptr) is true, a context will
never be inserted into the map, so check this first and return nullptr.

If the input value Codec != nullptr, it would always fail the checks in the else
branch.

avcodec_parameters_to_context() does not check for nullptr inputs, so ensure a
nullptr is not dereferenced.

Be explicit with comparisons to nullptr.

No functional change.