cmake: Fix cross-compilation to windows.
The change committed in 6012678e98 causes a fatal error if the ffmpeg
applications can't be found and used to determine its feature set.
This is fine for local compiles but cross compiled executable can't be
executed, so there's no point in requiring their existence (at this
point in the compilation).  Exempt cross compilation from this new
check.