cmake: Support the JAVA_HOME environment variable when building for android.
Building for android was only using the android studio installation of
java.  This worked fine for android studio installations that
contained Java 11 or Java 17, but failed at the very end for Java 21.
The problem is that the Qt androiddeployqt tool specifically invokes
gradle 7.2, and gradle 7.2 only supports up to Java 17.

Fix the build support to honor the JAVA_HOME environment variable so
that other version of java can be specified at build time.  The cmake
build instructions now look in the following places to find which java
to use:

    1) The environment variable JAVA_HOME
    2) The option variable MYTH_JAVA_HOME
    3) The android studio installation
    4) The system version

The build will now also check whether or not the version of java it is
using is compatible with androiddeployqt.