Fix several "may be used uninitialized" warning messages.
There are several functions that return an empty data structure as a
fallback at the end of the function.  These uninitialized warning
messages are only seen in release compiles because in a release
compile all assert() macros are compiled out.  Initialize these
structures to their default values to prevent this warning.