Minor rewrite to the recording extender test code.
The current code violates the C++ One Definition Rule where only a
single declaration of a class is allowed in a translation unit.  It
was originally written that way to try and minimize the amount of the
backend code that had to be linked in to make it compile.

Rewrite the code to use a derived class of "Scheduler" instead of a
second class named "Scheduler".  This satisfies the One Definition
Rule while also not requiring a cascading list of object files that
need to be linked in to the test code.