Fix potential null pointer dereferences in videodlg.cpp.
The compiler seems to believe that the m_d->m_videoList can be changed
inside of some of these functions.  Adding additional checks on the
validity of this pointer eliminates changes of dereferencing the
(possibly null) data pointer.

Unsure why putting the results of getListCache() into a temporary
variable eliminates the compiler warning about a null dereference,
unless somehow that forces the compiler to prune paths and prune out
the path that includes the nullptr.