tidy: Remove float->double->float conversion.
The sqrt() function operates on a double and returns a double, so the
current usage ends up implicitly promoting the supplied single
precision number to double precision number, squaring it, and then
implicitly converting the result back to a single precision number.
Using sqrtf() or std::sqrt() keeps the entire operation in the single
precision number space.