cmake: Add command to generate the compilation database.
By default, cmake generates a compilation database for each project
that it compiles.  In our case, that unfortunately means it generates
one database for the MythTV sub-project and another one for the
MythPlugins sub-project.  Add a top level command to concatenate these
two databases together to create a single database for the entire
source tree.