tidy: Fix "virtual method during destruction bypasses virtual dispatch" warning.
In the destructor of an object, any virtual function that is called is
(by the C++ specification) a reference to the function declared by
that class, not any possible overrides of the class function.  Add the
class name to the function call to make this explicit and to eliminate
the warning message.