Fix compiler warning about using an integer function on a float.
The clang compiler is pointing out that the abs() function (which
takes an int and returns and int) is being used in a context where the
argument and the result are expected to be floats.  Using the
std::abs() function instead will result in choosing the floating point
version of this function.