Fix potential null pointer dereferences in videodlg.cpp. (take 2)
The m_d->m_videoList can change inside of some of these functions
because it is of type simple_ref_ptr, and a "dereference" of that type
is actually a call to the ::get() function which can explicitly return
nullptr.  Adding additional checks on the validity of this pointer
eliminates changes of dereferencing the (possibly null) data pointer.