tidy: Fix warning about inconsistent function parameter names.
The declaration of the function ProcessAudioPacket uses "codecContext"
for the first argument, while the definition of the function uses
"ctx".  Change the name in the definition to "context".  This is
considered consistent with the declaration because "context" it is
part of the declared name "codecContext", and furthermore changing it
brings the function into line with all the other function definitions
in that file that use the name "context".