IPTV HLS Do not reset sequence number on EXT-X-DISCONTINUITY
The EXT-X-DISCONTINUITY tag indicates that there can be a change
in file format, number and identifiers of tracks and a bit more
as described in https://datatracker.ietf.org/doc/html/rfc8216#section-4.3.2.3
When the seqeunce number is reset then all segments described in the playlist
are played again, causing repeated reload and redisplay of segments that
have already been displayed. This is not something that should be done
when a EXT-X-DISCONTINUITY tag is seen.
Example stream is https://food-us.xumo.wurl.tv/playlist.m3u8
where the sections between the items are marked with
the EXT-X-DISCONTINUITY tag.
The new behavior is similar to that of VLC on the example stream.

Refs #936