EIT fixup for ITV HD multiplexes on Astra-2
Add EIT UK fixup processing for the multiplexes of the ITV HD
channels on the Astra-2 satellite, part of Freesat.
The fixup processing does modify the received EIT program guide
data to a format that is suitable for use in MythTV.
The EIT fixup is selected on a per-multiplex basis; the last
reshuffle moved the ITV HD channels to multiplexes 2113 and 2114
which were previously not selected for EIT fixup.

(cherry picked from commit 4954da3092e721149394e29242d6b4a2ed871924)