Support additional grouping options in program lister and finder.
Traditionally, MythTV has grouped programs by channel number and
callsign when displaying guide data in the EPG, program lister and
program finder.  This is done to unclutter the listings when the same
channel is available on multiple videosources.

This change adds support for the new grouping options listed below to
the program lister and program finder.  The EPG is not affected by
this change.  The traditional grouping by channel number and callsign
is called simply "Group By Channel Number".

Group By Call Sign: This option groups solely by callsign.  It is
useful when the same channel is available on multiple videosources but
with different channel numbers.

Group By Program ID: This option groups solely by program ID.  It is
useful when the same program is simulcast on multiple channels.  For
example, at present, many NBA and NHL games are simulcast on TNT and
TruTV.  In lineups that include SD and HD channels and/or Eastern and
Pacific channels, the can result in the same program being on up to 8
channels!  with this grouping option, those programs will only be
displayed once.

Group By None: This option doesn't perform any grouping at all.  It is
mainly included for completeness and for any rare cases where it might
prove useful.

The new grouping options are "sticky" in that they remain in effect
for all future guide displays until it is changed.  The
MythCenter-Wide theme has been updated to display the current grouping
option.  It is hoped that other them authors will also update their
themese.

Finally, this new grouping is deterministic.  This is in contrast to
the previous grouping which relied on MySQL's liberal grouping support
and could return non-deterministic results.  This new grouping chooses
programs using the following preferences: highest channel priority,
lowest videosource, lowest channel number.