Remove raw card IDs from a couple of user facing places.
As discussed on the mythtv-users mailing list, displaying long and
short, input names is preferred instead of card/input IDs.  This
commit removes the display of card IDs in a couple of places.  There
are probably still others but they can be fixed when noticed.

In the System Status // Input Status screen, the card IDs are removed
leaving the long input names as the primary input identifiers.

In any screens using the ProgramInfo::ToMap() function, the values of
the "card" and "input" map entries are swapped.  The "card" entry now
uses the short input name instead of card ID.  The obscure "input"
entry now uses the card ID for anyone who really, really, really still
wants to use card IDs for now.