IPTV HLS Support AES-128 encrypted channels
Support decoding of AES-128 encrypted channels.
This was already implemented a long time ago but needed
some fixing to make it work with today's streams.
Check existence of http(s) URLs by reading the first part
and do not start recording when the URL does not exist.
This prevents waiting when playing IPTV channels with Live TV.

Refs #936