IPTV HLS Restore check on QCoreApplication
Restore the check if we are running as a QCoreApplication
before starting a download as part of the GuessProtocol function.
This is needed to be able to run the code as part of the
unit tests which are apparently not QCoreApplications.
Note that even with this check restored the test does
still fail. This is because a http or https URL is now
only considered valid if a download from that URL is
possible, instead of only being syntactically correct.

Refs #936