IPTV HLS Restore GuessProtocol old behavior
Restore the old behavior of the GuessProtocol function.
This means that http/https URL's that cannot be downloaded
are identified as http_ts type of files.
This fixes the unit test TestIPTVRecorder.
The underlying problem in iptvtuningdata.h is that
the enum IPTVProtocol does mix values representing
a protocol, e.g. udp and rtp, with values representing
the content of what is downloaded.
In this case http and https are protocols and should
be part of the enumeration.
In contrast, http_ts specifies a TS stream in a https stream
and http_hls specifies a HLS playlist in a http or https stream.
The protocol specifications are considered to be valid
when they are syntactically correct, even when they do not exist.
The content can only be verified if the URL resource can
be downloaded.
There should be separate enumerations for protocol and
for resource content. This is likely to be non-trivial
and hence the old behavior is restored for now.

Refs #936