Fix SatIP tuner/frontend identification
Fix SatIP tuner identification when the SatIP server (minisatip)
has more than one type of tuner, e.g. both satellite and cable
or digital terrestrial.
The tuner identification is used to generate the "fe=xx" parameter
of the tuning URL. Although the use of the "fe=xx" parameter
is not recommended in the SatIP specification it is necessary
to fully support the same features as the /dev/dvb/*/* tuners.

Refs #1036