Remove hardware profiling dependence on python3.future.
Fedora 41 has stopped packaging this python module, and others may
follow suit.  The only dependency on this package is the old_div
function which provides python2 style division (either float or
integer division depending on the arguments).  All five references to
this function appear to be using floating point division, so calls to
this function can be replaced with the standard python3 division
operator '/'.