Simplify pointer manipulation in zmserver.cpp.
Clang-tidy pointed out this file as casting integers to pointers,
which conceals information from the optimizer if the integer was
originally cast from a pointer.

The pointer in question has an alignment of 1, so the conversion into
the integer domain and back for the addition is entirely unnecessary.
Simply add the necessary offset to get the desired alignment.  Its
highly unlikely that these additions will ever be called since the
pointer is a result of mmap() and kernel pages have 1024 byte aligned.