Fix building on OSX with the most recent version of mesa.
Older mesa wasn't detected by configure on OSX, and therefore MythTV
only built in support for VideoToolbox.  The latest mesa is detected
and MythTV tries to build support for both VideoToolbox and OpenGL,
which fails.  No idea why this changed, but the cause of the compile
failures is that EGLNativeDisplayType is defined as "void*" on most
platforms, but its defined as "int" on Apple.  This conflicts with
MythTV's hard coded signature for the EGLGETPLATFORMDISPLAY function.
Providing a variant signature for Apple solves the compilation
problem.