Initial support for replacing QWebKit with QWebEngine
Very lightly tested on Qt 5.15.17 and 6.6.2 and only using MythBrowser which
is the primary reason for it anyway. It's not yet completely equivalent to
the original some features like download support and the browser API are
missing.

There is a bug in MythUI when using Qt6 and running MythTV in a window any
scaling factor is not applied properly to the GUI but is to the window size.

Don't know what needs to be done for cmake so that will probably be broken.