simplify MPoolThread::run()
MPoolThread::m_lock is always locked except during the m_wait.wait() call
and the calls to m_pool.NotifyAvailable() and m_pool.NotifyDone().

If m_doRun is false at the end of the infinite loop (only possible if
MPoolThread::SetRunnable() and then MPoolThread::Shutdown() are called before
MPoolThread::run() acquires the lock, since m_doRun will be false and
m_runnable will not be nullptr when m_runnable is tested in the loop) and the
else branch is removed, the next iteration will skip the wait and perform the
same operations since the lock is never released.  Also, m_doRun is never set
back to true once it is false and m_runnable is never set to a non-nullptr
value if m_doRun is false.

This reveals m_runnable != nullptr is the true loop condition (and also that
the QRunnable will be run after Shutdown() is called).

The test for m_doRun is preserved so m_pool.NotifyAvailable() is still not
called when m_doRun is false.

The code is then reordered so the loop condition can be used.

Since the lock was released and another thread could have been waiting in
MPoolThread::SetRunnable() or MPoolThread::Shutdown(), we need to check
m_runnable == nullptr in addition to m_doRun before waiting.