Prevent deadlock in MThreadPool
It deadlocks if any thread is running and m_reserved is true since the thread
will call m_pool.ReleaseThread() which will attempt to lock
MThreadPool::m_priv::m_lock, but MThreadPool::Stop() holds that lock and is
waiting to acquire the lock of MPoolThread::m_lock in MPoolThread::Shutdown()
which was not released.

This deadlock can be triggered in mythfrontend by using
http://localhost:6551/MythFE/GetRemote and then exiting within 10 seconds.
The shutdown is still delayed by the 10 seconds that HttpWorker::run() is
waiting.