tidy: Fix potential memory leak warning.
The sendKeyPress() function may or may not send the passed message,
depending on the contents of the member variable "children".  If it
doesn't send the message there is a memory leak.  If it sends the
message more than once, the behavior is undefined because postEvent()
takes ownership of the variable.  Change the code to create a new
event immediately before each call to to postEvent().