Prepare for converting from USING_XXX to CONFIG_xxx defines.
CONFIG_XXX defines are nicer because they only have to be defined once
in mythconfig.h, not redefined in every source code directory where
they are used.  Also, an "#if CONFIG_XXX" statement prints a warning
if the variable doesn't exist, whereas an "ifdef USING_XXX" statement
doesn't print anything if the variable doesn't exist.  Did the build
system not get updated to add the #define in that directory, or the
feature truly not needed in that directory?  Unless there's a
compilation failure, there's no indication that code was/wasn't
compiled when it should have been.

1) Add back in the CONFIG_XXX variables what were stripped out of
mythconfig.h.in when it was created, because they weren't currently
used in any of the cpp files.

2) Add a couple of new CONFIG_XXX variables to configure/mythconfig.h.

3) Clean up the doxygen documentation in a single commit.