Quiet cmake 4.0 warning about cmp0135.
Cmake 4.0 prints a warning about this cmake policy being set to use
its old behavior.  Change the code so that if this policy exists, it
uses the "new" behavior but also sets the DOWNLOAD_EXTRACT_TIMESTAMP
keyword on the three external projects that need it.  The keyword
can't just always be set, because it doesn't exist prior to the
release that introduced this policy.