Web App: Add Previously Recorded Sort options, plus bug fix
Add sort option to all columns of Previously Recorded, except Time,
which is sorted in conjunction with Date.

Sort on Status sorts on the numeric values, not the translated text
values, so you get the same status grouped together, but not alphabetic.

Sort on Status only sorts on the second part of the Type -- Status
combination (the Status).

Sort on Status sorts Re-Record and Recorded together. This is a
technical limitation.

Fix bug in Previously Recorded, Recorded and Upcoming, where updates
done in a pert of the list after scrolling down, may not be visible,
even after a refresh.

Refs #1068