Work around a timeout and eventual playback exit during fast-forward/rewind.
Something changed, I think around the time of the ffmpeg 7.1 update,
that causes fast-forward/rewind to sometimes hang waiting for decoded
video buffers.  I suspect an error in the video buffer life cycle but
I've so far been unable to pin it down.  For the time being, work
around the problem by clearing the render one flag when we warn about
waiting too long for buffers.