osx: Fix videotoolbox test and set CONFIG_VIDEOTOOLBOX properly.
The cmake test for whether to disable Apple videotoolbox (modeled off
the configure test) needs to be fixed.  The videotoolbox specific
items are listed in the FFmpeg encoders, not the FFmpeg decoders.

Also set the CONFIG_VIDEOTOOLBOX variable once its determined whether
or not MythTV should be using the videotoolbox code.