DBus screensaver always report being awake
Always return false for the Asleep function of the DBus screensaver.
Remove the inhibite state bookkeeping because that is not used anymore.
The Asleep function is only relevant for the X11 screensaver.
The value of this function is used in mythinputdevicehandler.cpp to discard input events when the screen is powered off. This is only useful with the X11 screensaver; the DBus screensaver has the state "inhibited" and "not inhibited" but it does not say anything about the screen actually being saved or not.

Fixes #1064

(cherry picked from commit a52127704d7c80a298a803cd82e2435466742c01)