Service API Authentication support
Add support for username and password authentication for the backend
service API. New methods Myth/LoginUser and Myth/GetUsers. Authenticate
using an access token supplied by the Myth/LoginUser method following
a valid user name and password.

A new setting APIAuthReqd determines whether authentication is required
and whether it is required only for remote IP addresses or for all IP
addresses. By default it is set to NONE.

All methods are protected when APIAuthReqd is enabled except for
Myth/LoginUser and Myth/GetConnectionInfo.

This builds on the existing mythsession framework already developed for
the defunct web frontend. The admin user is already pre-populated in
the database.