Web App Authentication Support
Add support for username and password authentication for the backend
web app.

Add a new section "Users" under Dashboard -> Settings. This supports
enabling the authentication requirement, selecting whether all IP
addresses or only remote IP addresses are protected, creating and
maintaining user login names and passwords.

Add a login link to the top banner of the web app. Also force a login
prompt if the backend requires a login (via http 401 response).

Update all communications to the backend to include the access token
from login if a login was performed.