Convert "do" loops to "while" loops. (mythfilldatabase)
Using a "while" loop is easier to understand because the exit
condition of the loop is stated up front.  In one of these cases, the
exit condition was pulled 360(!) lines earlier than its previous
position.  Each of these while loops is reached at the start element
of a given tag, so the first clause in the while loop will always be
false and like the prior "do" statements they will all be entered and
processed.  (None of these are empty elements.)

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines/avoid-do-while.html