Convert "do" loops to "while" loops. (libmythbase)
Using a "while" loop is easier to understand because the exit
condition of the loop is stated up front.

In one function it makes more sense to use a "for" loop than a "do" or
a "while" loop.

In another function. don't put multiple calls to query.next() in the
same function.  Use a single while loop on the result of that function
call to process all the returned data.

https://clang.llvm.org/extra/clang-tidy/checks/cppcoreguidelines/avoid-do-while.html