Convert "do" loops to "while" loops. (libmythtv/programdata.cpp)
Don't put multiple calls to query.next() in the same function.  Use a
single "while" loop on the result of that function call to process all
the returned data.  Events that need to know if there were no results
can test if the position is still QSql::BeforeFirstRow after the
processing loop.