Close memory leak in loglevelAdd()
The verboseMap and loglevelMap entities are of
type QMap. When the values inserted into a QMap
are from allocated memory, the recommended way to
clean up, is to invoke
  qDeleteAll( map ); // delete all values in map
  map.clear(); // remove map references to values

In verboseAdd(), new VerboseDef is invoked.
In loglevelAdd(), new LoglevelDef is invoked.

To properly clean up without a memory leak,
in logStop() and verboseInit() we need to
call qDeleteAll() and clear().

Resolves #1104