Close memory leak in getMythSortHelper
The syntax being used for std::make_shared<>
was improper. The whole point of using this
service is to automatically allocate memory
and invoke the constructor. By eliminating
the unnecessary 'new' call, we remove the
memory leak. In addition we simplify the
invocation by removing the unnecessary
constructor call. The correct syntax makes
it safer, faster, and less leaky.

Resolves #1106