Close memory leaks in PulseAudio output handler
Adding a check that m_mainloop is non-null before
invoking pa_context_drain(). It is extremely
unlikely that such a circumstance will ever happen.
The lock is only aquired if it is non-null,
and the code for handling pa_context_drain()
requires that the lock be held before calling
pa_threaded_mainloop_wait(), so this small
modification makes the code safer.

Resolves #1100