Close memory leak in StandardSetting
The destructor for StandardSetting released memory
for m_children and m_targets members, but failed
to release the memory for m_storage. Each class
using StandardSetting as a base uses 'new' to
create storage, and the constructor assigns
that allocated memory to m_storage.

This modification adds code to the destructor
to delete m_storage. This closes the memory leak.

Resolves #1120