Fix New, Delete type mismatch
At the allocation point a MythVideoTextureOpenGL
class was created. At the deletion point, a
MythGLTexture was released. MythGLTexture is a
base class for MythVideoTextureOpenGL.

If you intend to delete a derived class object
through a base class pointer, the base class must
have a virtual destructor. This ensures that the
correct destructor for the derived class is
called when the object is deleted.

This modification stipulates the destructor for
MythGLTexture to be virtual. This ensures that
the type upon deletion matches the type upon
creation.

Resolves #1121