Stop memory use after free (#1132)
MConcurrent was used to launch a separate thread to
take on the task of saving the playback position.
However, MConcurrent doesn't provide any way of
tracking the progress of that task. This is now
switched to use MThreadPool. With a handle to this
thread pool, we can call waitForDone() to wait
for the completion of the saving playback position
task in TV::~TV(). This eliminates the race
condition between the threads, and stops the
reference to memory which has already been freed.

Resolves: #1131