mythcommflag d2: Fix potential crash in CannyEdgeDetector.cpp.
The resetBuffers function is called for every frame to release the
buffers from the previous frame and allocate new buffers for the
current frame.  As an optimization, if the previous frame and the new
frame are the same size, it reuses the previous buffers.  This works
great, unless the previous call to the function couldn't allocate a
buffer and returned an error.  The current invocation code will then
return success even though all the buffers are null.

Update the function so that comparing sizes short cut only occurs if
the buffers currently point to valid memory.