Safer close of memory leak in StandardSetting
Boolean 'm_newdStorage' has been added to the
StandardSetting class. This will be set to true
when 'new' was used to allocate the memory which
is being put into Storage. In the destructor for
StandardSetting, a test has been added so that
such memory will only be deleted if m_newdStorage
is true. This allows us to free allocated memory
associated with StandardSetting, without trying
to free memory which wasn't allocated from the heap.

All of the classes derived from StandardSetting which
use 'new' to allocate the stored contents have been
updated to set 'm_newdStorage' to true. This will
ensure that upon destruction, that memory will be
freed.

Resoves: #1120