use AudioOutput::DecodeAudio() instead of AudioOutputUtil::DecodeAudio()
The only difference is the AVFrame AudioOutput::m_frame will be reused instead
of allocating and freeing an AVFrame with every call and the samples will be
converted to FORMAT_S16 if AudioOutput::CanProcess() returns false.  However,
FORMAT_S16 is the desired format and no conversion was performed, so the
audio samples were assumed to already be FORMAT_S16.