Close memory leak with smart pointers
All of the classes derived from StandardSetting which
used to use 'new' to allocate memory have been
switched to use std::make_shared<> to create an
instance with a smart pointer. When the last reference
to such a smart pointer is destructed, the memory
is freed. This eliminates StandardSetting related
memory leaks.

Resolves: #1120