Close memory leaks by using destructors
All of the classes derived from StandardSetting which
use 'new' to allocate memory have been updated to add
destructors which call "delete GetStorage();" This
frees up memory allocated from the heap when an
instance of that class falls out of scope or is
deleted. This closes the StandardSetting memory
leaks.

Resolves: #1120